requirements.txt
aiohttp~=3.10.10
azure-eventhub~=5.12.2
psutil~=6.1.0